+++ flutter_app/first_flutter_app/lib/cubit/app_cubit.dart
@@ -1,36 +1,27 @@
 import 'package:english_words/english_words.dart';
+import 'package:first_flutter_app/cubit/app_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class AppState {
-  var current = WordPair.random();
-  var favorites = <WordPair>[];
-
-  copyWith({current, favorites}){
-    var newState = AppState();
-    newState.current = current ?? this.current;
-    newState.favorites = favorites ?? this.favorites;
-    return newState;
-  }
-}
-
 class AppCubit extends Cubit<AppState> {
-  AppCubit(super.initialState);
+  AppCubit(): super(AppState(current: WordPair.random(), favorites: []));
 
   void getNext() {
     emit(state.copyWith(current: WordPair.random()));
   }
 
   void toggleFavorite() {
-    if (state.favorites.contains(state.current)) {
-      state.favorites.remove(state.current);
-    } else {
-      state.favorites.add(state.current);
+    List<WordPair> favorites = List.from(state.favorites);
+    if (favorites.contains(state.current)) {
+      favorites.remove(state.current);
+    } 
+    else {
+      favorites.add(state.current);
     }
-    emit(state.copyWith());
+    emit(state.copyWith(favorites: favorites));
   }
 
   void removeFavorite(WordPair wordPair) {
     state.favorites.remove(wordPair);
-    emit(state.copyWith());
+    emit(state.copyWith(favorites: state.favorites));
   }
 }
