+++ flutter_app/first_flutter_app/lib/page/favorite_page.dart
@@ -1,18 +1,19 @@
 import 'package:first_flutter_app/cubit/app_cubit.dart';
+import 'package:first_flutter_app/cubit/app_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class FavortitePage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    var cubit = context.read<AppCubit>();
     return BlocBuilder<AppCubit, AppState>(
+      buildWhen: (previous, current) =>
+          previous.favorites != current.favorites,
       builder: (context, state) {
-
         if (state.favorites.isEmpty) {
           return Center(child: Text("Your favorite list is empty"));
         }
-
+    
         return Padding(
           padding: const EdgeInsets.all(32.0),
           child: Column(
@@ -28,7 +29,8 @@ class FavortitePage extends StatelessWidget {
                         Spacer(),
                         IconButton(
                             onPressed: () {
-                              cubit
+                              context
+                                  .read<AppCubit>()
                                   .removeFavorite(state.favorites[index]);
                             },
                             icon: Icon(Icons.delete))
