+++ flutter_app/first_flutter_app/lib/page/generator_page.dart
@@ -1,4 +1,5 @@
 import 'package:first_flutter_app/cubit/app_cubit.dart';
+import 'package:first_flutter_app/cubit/app_state.dart';
 import 'package:first_flutter_app/widget/big_card_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -6,10 +7,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 class GeneratorPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-
-    var cubit = context.read<AppCubit>();
-
     return BlocBuilder<AppCubit, AppState>(
+      buildWhen: (previous, current) {
+        return (previous.favorites != current.favorites) || (previous.current != current.current);
+      },
       builder: (context, state) {
         IconData icon;
         if (state.favorites.contains(state.current)) {
@@ -29,7 +30,7 @@ class GeneratorPage extends StatelessWidget {
                 children: [
                   ElevatedButton.icon(
                     onPressed: () {
-                      cubit.toggleFavorite();
+                      context.read<AppCubit>().toggleFavorite();
                     },
                     icon: Icon(icon),
                     label: Text('Like'),
@@ -37,7 +38,7 @@ class GeneratorPage extends StatelessWidget {
                   SizedBox(width: 10),
                   ElevatedButton(
                     onPressed: () {
-                      cubit.getNext();
+                      context.read<AppCubit>().getNext();
                     },
                     child: Text('Next'),
                   ),
